+++ sandbox-backend/src/main/java/com/exadel/sandbox/team5/configs/PasswordEncoderTest.java
@@ -1,0 +1,15 @@
+package com.exadel.sandbox.team5.configs;
+
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+public class PasswordEncoderTest implements PasswordEncoder {
+    @Override
+    public String encode(CharSequence charSequence) {
+        return charSequence.toString();
+    }
+
+    @Override
+    public boolean matches(CharSequence charSequence, String s) {
+        return charSequence.toString().equals(s);
+    }
+}
